add etcd watcher for blockaffinity
